These tests/features now require libtiff
